rearrange packages
