--- src/main/java/Storage/AccountsStorage.java
@@ -1,17 +1,0 @@
-package Storage;
-
-public class AccountsStorage extends Storage{
-    public AccountsStorage() {
-        super("storage/accounts.txt");
-    }
-
-    @Override
-    public void writefile(String text) {
-        super.writefile(text);
-    }
-
-    @Override
-    public String readfile() {
-        return super.readfile();
-    }
-}
