--- src/main/java/Storage/MessagesStorage.java
@@ -1,17 +1,0 @@
-package Storage;
-
-public class MessagesStorage extends Storage{
-    public MessagesStorage(String name) {
-        super("storage/messages/" + name + ".txt");
-    }
-
-    @Override
-    public void writefile(String text) {
-        super.writefile(text);
-    }
-
-    @Override
-    public String readfile() {
-        return super.readfile();
-    }
-}
