--- src/main/java/Storage/MessagesStorageGateway.java
@@ -1,5 +1,0 @@
-package Storage;
-
-public interface MessagesStorageGateway {
-
-}
