--- src/main/java/Storage/MessagesStorageUsecase.java
@@ -1,4 +1,0 @@
-package Storage;
-
-public class MessagesStorageUsecase {
-}
